Move some functions into path_matcher.cc
